Re structure now organize everything in place
