+++ src/components/game/Speedburst.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Position } from '../types';
+import { Position } from '../../types';
 
 export default function SpeedBurst({ position }: { position: Position }) {
   return (
