+++ src/components/ui/Score.tsx
